Open connections when used; update connections when needed
Previously we created all the connections up front when the application starts.
This causes problems when the connection is smb which may not have the share
information until the messages come in on the kafka queue. This moves all the
connections to be lazily created.

This also can get a notification from connmgr when a new/deleted/modified connection
happens. If we detect an existing connection, we update it right away before processing
any more records.

We also now detect when policyengine sends a STOP signal which means we may need to ignore
current records on the kafka queue. These records will be continuously skipped until we process
some records and can commit on the queue.